Extract PluginCoordinator from MainController
Move the Plugin support feature into a new ui/PluginCoordinator via the
CoordinatorHost pattern — the largest single step of the MainController
decomposition.

PluginCoordinator owns the per-window PluginRegistry/PluginInstaller, the
started-plugin instances, the plugin-contributed tool windows + editor menu
items, the "Browse Plugins" picker, and the two inner adapters
(PluginContextImpl/ActiveEditorImpl). It covers the whole flow:
- applyPlugins: declarative snippet/template source dirs + keymap overrides +
  plugin.<id>.<cmd> external commands, then each Java plugin's start(ctx)
- the PluginContext adapter (registerCommand/bindKey/registerToolWindow/
  addEditorMenuItem/addStatusBarSegment/activeEditor/...)
- browse/install-from-disk/install-from-registry/uninstall + the capability
  -disclosure enable-confirm, and the runtime gates

The shared PluginManager (owned by WindowManager, classes load once) is passed
in. The per-window collaborators (CommandRegistry/KeymapManager/SnippetManager/
TemplateRegistry/ToolWindowManager/StatusBar/SettingsWindow/ConfigManager) are
passed to the constructor; a tiny Ops covers openPath + the shared Git error
dialog. MainController keeps the pluginManager field + thin disposePlugins()
(the WindowManager window-close contract) and pluginsEnabled() delegates, and
wires applyPlugins / gateToolWindows / editorMenuItems / the plugins.* +
view.togglePlugins commands to the coordinator. The buffer-gating call is
null-guarded (the first setupToolWindows pass runs before the coordinator is
built — it needs toolWindows — and no plugin tool windows exist until
applyPlugins; the FX-harness window-build tests caught this).
PluginCoordinator.pluginCapabilitySummary is the public static the Settings
page now calls.

Behavior-preserving — bodies moved verbatim (MainController.this.X -> host.X/
ops.X; the collaborators referenced as coordinator fields).

MainController: 9578 -> ~9004 lines (-574). New PluginCoordinator ~700 lines.
Verified: full `mvn verify` green (spotless + JaCoCo + all tests incl. the 8
FX-harness window-build tests that construct the coordinator during init), and
device-tested live via the MCP execute_command bridge — the example plugin's
Java-registered command (plugin.example.insertStamp), its tool-window toggle
(plugin.example.hello), and its declarative external command
(plugin.example.date) all loaded + ran through the moved PluginContextImpl.start
path.
