+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **Internal: extracted `PluginCoordinator` from `MainController`.** The Plugin support feature — discovery/apply/install/uninstall plus the per-window `PluginContext` adapter that lets a plugin contribute commands, keybindings, tool windows, status-bar segments, and editor menu items — moved into its own `CoordinatorHost`-style coordinator. It owns the per-window `PluginRegistry`/`PluginInstaller`, the started-plugin instances, the plugin-contributed tool windows + editor menu items, the "Browse Plugins" picker, and the two inner adapters (`PluginContextImpl`/`ActiveEditorImpl`). The shared `PluginManager` (owned by `WindowManager`) is passed in; `MainController` keeps the `pluginManager` field + thin `disposePlugins()`/`pluginsEnabled()` delegates and wires `applyPlugins` / the tool-window gating / the editor-menu contributor / the `plugins.*`+`view.togglePlugins` commands to the coordinator. `MainController` shed ~574 lines (9,578 → ~9,004). Device-tested live (MCP bridge): the example plugin's Java-registered command, its tool-window toggle, and its declarative external command all loaded + ran through the moved `PluginContextImpl.start` path. No behavior change — part of the ongoing `MainController` decomposition (the largest single extraction in the campaign).
 - **Internal: extracted `HistoryCoordinator` from `MainController`.** The Local File History feature (per-file save/label/delete snapshots stored outside the file) moved into its own `CoordinatorHost`-style coordinator: it owns the `HistoryService` (off-thread snapshot/blob engine) + the `FileHistoryPanel`, and the whole flow — record (save / autosave / external / pre-delete), the per-file + folder-history views, label / recent-changes pickers, and revision diff/restore (reusing the shared `DiffCoordinator`). `MainController` keeps the File History `ToolWindow` (built with the coordinator's panel) and wires the record hooks + the `history.*`/`tool.fileHistory` commands + the project-tree "Show Local History" item to it. `MainController` shed ~300 lines (9,881 → ~9,578). Device-tested live (MCP bridge): saving a file recorded a revision into `history/index.json`, and the history commands ran clean. No behavior change — part of the ongoing `MainController` decomposition.
 - **Internal: moved the Git operations into `GitCoordinator`.** The user-facing Git mutation operations that were still in `MainController` — `gitOp`/`gitCommit`/`discardChanges`, branch checkout/create, fetch/pull/push (`gitSync`), the stash commands, the active-file stage/unstage/discard, the commit-focus action, and the `gitError` dialog — moved into `GitCoordinator` (which already owned the `GitService` + repo state), so the whole Git engine + its mutations now live in one cohesive coordinator. `MainController` keeps the panels/tool-windows/branch-popup + command registrations (delegating to `git.*`). `MainController` shed ~294 lines (10,175 → ~9,881). No behavior change. (The commit/branch/log/blame/clone UI wiring is a follow-up.)
 - **Internal: extracted `DiffCoordinator` from `MainController`.** The diff + merge-conflict viewer (the `DiffService`, the diff-tab open/refresh machinery, the "apply change" hunk flow through an undoable buffer with Undo/Save, the compare entry points — vs HEAD / vs commit / compare-with-file / Git-panel rows / a commit's file — patch export, and merge-conflict resolution) moved into its own `CoordinatorHost`-style coordinator. Git-backed diffs reach the repo via the shared `GitCoordinator`; `openDiff` + the `DiffSide` interface stay package-visible so the Local File History flow reuses them. `MainController` shed ~317 lines (10,492 → ~10,175). No behavior change — continuing the `MainController` decomposition.
