+++ src/main/java/com/editora/ui/MainController.java
@@ -199,24 +199,8 @@ public class MainController implements com.editora.mcp.McpBridge {
     private com.editora.template.TemplateRegistry templates;
     /** Shared across windows (owned by WindowManager); plugin classes load once, instances are per-window. */
     private com.editora.plugin.PluginManager pluginManager;
-    /** Java plugin instances started in this window (for {@code stop()} on window close). */
-    private final java.util.List<com.editora.plugin.Plugin> startedPlugins = new java.util.ArrayList<>();
-    /** Editor right-click items contributed by plugins in this window (label + action over the editor). */
-    private final java.util.List<EditorMenuContribution> pluginMenuItems = new java.util.ArrayList<>();
-    /** Fetches the remote plugin-registry index (created in init when a PluginManager is present). */
-    private com.editora.plugin.PluginRegistry pluginRegistry;
-    /** Downloads/verifies/unpacks plugin zips (created in init when a PluginManager is present). */
-    private com.editora.plugin.PluginInstaller pluginInstaller;
-    /** The "Browse Plugins" picker (in-scene overlay); items come from {@link #browseEntries}. */
-    private QuickOpen<com.editora.plugin.RegistryEntry> browsePalette;
-    /** The last-fetched registry entries, shown by {@link #browsePalette}. */
-    private java.util.List<com.editora.plugin.RegistryEntry> browseEntries = java.util.List.of();
-    /** Whether the last-fetched registry index verified against the bundled signing key. */
-    private boolean browseSigned;
-
-    /** A plugin-contributed editor menu item: a label + an action over the {@link com.editora.plugin.ActiveEditor}. */
-    private record EditorMenuContribution(
-            String label, java.util.function.Consumer<com.editora.plugin.ActiveEditor> action) {}
+    /** Plugin support (discovery/apply/install + the per-window PluginContext); see {@link PluginCoordinator}. */
+    private PluginCoordinator pluginCoordinator;
 
     private com.editora.completion.CompletionEngine completion;
     private FileFinder fileFinder;
@@ -320,12 +304,6 @@ public class MainController implements com.editora.mcp.McpBridge {
     private GitLogPanel gitLogPanel;
     private GitLogPanel.Actions gitLogOps; // reused by the git.log.* palette commands (act on the selected commit)
     private ToolWindow gitLogToolWindow;
-    /**
-     * Tool windows contributed by plugins. They act on the active editor (via {@code ActiveEditor}), so —
-     * like the built-in buffer tool windows — their stripe button is gated on an open buffer in
-     * {@link #updateBufferToolWindows()} (a non-buffer tab like Welcome has nothing for them to act on).
-     */
-    private final java.util.List<ToolWindow> pluginToolWindows = new java.util.ArrayList<>();
     /** The path the Git Log is currently filtered to (file history), or null for the whole repo. */
     private Path gitLogFilter;
     /** Local File History: snapshots local files on save/auto-save/external reload (off-thread). */
@@ -513,7 +491,20 @@ public class MainController implements com.editora.mcp.McpBridge {
                 this::exportConfig,
                 this::showDebugLog);
         this.settingsWindow.setPluginManager(pluginManager); // shared; lists discovered plugins on the Plugins page
-        this.settingsWindow.setPluginActions(this::browsePlugins, this::installPluginFromDisk, this::uninstallPlugin);
+        this.pluginCoordinator = new PluginCoordinator(
+                coordinatorHost,
+                registry,
+                keymap,
+                snippets,
+                templates,
+                toolWindows,
+                statusBar,
+                settingsWindow,
+                config,
+                pluginManager,
+                pluginOps());
+        this.settingsWindow.setPluginActions(
+                pluginCoordinator::browse, pluginCoordinator::installFromDisk, pluginCoordinator::uninstall);
         this.settingsWindow.setDictionaryActions(
                 this::openTechnicalDictionary, this::openPersonalDictionary); // Spell Check file links
         this.settingsWindow.setSnippetManager(snippets); // backs the Settings → Snippets management page
@@ -562,7 +553,8 @@ public class MainController implements com.editora.mcp.McpBridge {
         setupRecentFiles();
         setupJumpPickers();
         setupProjects();
-        applyPlugins(); // register plugin commands/tool windows/hooks (before restore, so their visibility restores)
+        pluginCoordinator
+                .applyPlugins(); // register plugin commands/tool windows/hooks (before restore, so visibility restores)
         toolWindows.restore();
         // Honor a persisted Zen state on launch: the view options + chrome already read the flag via
         // the apply paths; this hides the side stripes (restore() opened nothing — windows were
@@ -603,11 +595,7 @@ public class MainController implements com.editora.mcp.McpBridge {
 
     /** Injected by {@link WindowManager} before {@link #init}; shared across windows (classes load once). */
     void setPluginManager(com.editora.plugin.PluginManager pm) {
-        this.pluginManager = pm;
-        if (pm != null) {
-            this.pluginRegistry = new com.editora.plugin.PluginRegistry();
-            this.pluginInstaller = new com.editora.plugin.PluginInstaller(pm);
-        }
+        this.pluginManager = pm; // the per-window PluginRegistry/PluginInstaller are built in PluginCoordinator
     }
 
     /** Whether plugins may load — the master gate (also off in Simple UI mode). */
@@ -615,558 +603,26 @@ public class MainController implements com.editora.mcp.McpBridge {
         return pluginManager != null && config.getSettings().isPluginSupport() && !simpleModeActive();
     }
 
-    /**
-     * Applies every enabled, error-free plugin to <em>this</em> window: declarative snippet/template source
-     * dirs + keymap bindings + external-command palette entries, then each Java plugin's {@code start(ctx)}.
-     * Runs once during {@link #init} (off no hot path). One {@code try/catch} per plugin → DebugLog, so a
-     * misbehaving plugin never breaks the window.
-     */
-    private void applyPlugins() {
-        if (!pluginsEnabled()) {
-            return;
-        }
-        boolean addedAssets = false;
-        for (com.editora.plugin.PluginDescriptor d : pluginManager.descriptors()) {
-            if (!d.enabled() || d.loadError() != null) {
-                continue;
-            }
-            try {
-                // Declarative: snippet/template source dirs (per-window registries) + keymap (shared).
-                java.nio.file.Path snDir = d.dir().resolve("snippets");
-                if (java.nio.file.Files.isDirectory(snDir)) {
-                    snippets.addExtraSourceDir(snDir);
-                    addedAssets = true;
-                }
-                java.nio.file.Path tplDir = d.dir().resolve("templates");
-                if (java.nio.file.Files.isDirectory(tplDir)) {
-                    templates.addExtraSourceDir(tplDir);
-                    addedAssets = true;
-                }
-                if (d.manifest().keymap != null && !d.manifest().keymap.isEmpty()) {
-                    keymap.applyOverrides(d.manifest().keymap); // shared keymap → applies to every window
-                }
-                // Declarative external commands → palette commands.
-                for (com.editora.plugin.PluginManifest.DeclaredCommand c : d.manifest().commands) {
-                    if (c == null || c.id == null || c.id.isBlank()) {
-                        continue;
-                    }
-                    String cid = "plugin." + d.id() + "." + c.id;
-                    String title = c.title == null || c.title.isBlank() ? cid : c.title;
-                    registry.register(com.editora.command.Command.of(cid, title, () -> runDeclaredCommand(d, c)));
-                }
-                // Java plugin.
-                if (d.hasJavaEntry() && d.classLoader() != null) {
-                    com.editora.plugin.Plugin plugin = pluginManager.instantiate(d);
-                    if (plugin != null) {
-                        plugin.start(new PluginContextImpl(d));
-                        startedPlugins.add(plugin);
-                    }
-                }
-            } catch (Throwable t) {
-                LOG.log(java.util.logging.Level.WARNING, "Plugin " + d.id() + " failed to apply", t);
-                setStatus(tr("status.plugins.failed", d.id()));
-            }
-        }
-        if (addedAssets) {
-            snippets.reload();
-            templates.reload();
-        }
-    }
-
-    /** Stops this window's plugins (window close). Each {@code stop()} is isolated. */
+    /** Stops this window's plugins (window close); WindowManager calls this by name. */
     void disposePlugins() {
-        for (com.editora.plugin.Plugin p : startedPlugins) {
-            try {
-                p.stop();
-            } catch (Throwable t) {
-                LOG.log(java.util.logging.Level.WARNING, "Plugin stop failed", t);
-            }
-        }
-        startedPlugins.clear();
-    }
-
-    // --- plugin registry (browse / install) ------------------------------------------------------
-
-    /**
-     * Fetches the configured registry index off-thread and shows the "Browse Plugins" picker. No-op (with a
-     * status hint) when plugins are disabled or no registry URL is set.
-     */
-    private void browsePlugins() {
-        if (!pluginsEnabled() || pluginRegistry == null) {
-            setStatus(tr("status.plugins.disabled"));
-            return;
-        }
-        String url = config.getSettings().getPluginRegistryUrl();
-        setStatus(tr("status.plugins.fetching"));
-        boolean requireSig = config.getSettings().isPluginRequireSignature();
-        pluginRegistry.fetch(url, r -> {
-            if (!r.ok()) {
-                setStatus(tr("status.plugins.fetchFailed", r.error()));
-                return;
-            }
-            // Signature gate: when "require signed plugins" is on, refuse an unsigned/unverified registry.
-            if (requireSig && !r.signed()) {
-                git.gitError(tr("status.plugins.unsigned"), tr("dialog.plugins.unsignedDetail"));
-                return;
-            }
-            browseSigned = r.signed();
-            browseEntries = r.entries();
-            if (browseEntries.isEmpty()) {
-                setStatus(tr("status.plugins.empty"));
-                return;
-            }
-            if (!browseSigned) {
-                setStatus(tr("status.plugins.unsignedAllowed"));
-            }
-            browsePalette.show(stage);
-        });
-    }
-
-    /** The picker label: "Name  version — Installed/Update available/Install/Requires Editora ≥ x". */
-    private String browseEntryLabel(com.editora.plugin.RegistryEntry e) {
-        String ver = e.version == null ? "" : e.version;
-        return (e.name == null || e.name.isBlank() ? e.id : e.name) + (ver.isBlank() ? "" : "  " + ver) + " — "
-                + browseEntryStatus(e);
-    }
-
-    /** Installed / Update available / Install / Requires-newer, comparing to the installed descriptor. */
-    private String browseEntryStatus(com.editora.plugin.RegistryEntry e) {
-        if (!meetsMinEditora(e)) {
-            return tr("plugins.status.requiresNewer", e.minEditoraVersion);
-        }
-        String installed = installedVersion(e.id);
-        if (installed == null) {
-            return tr("plugins.status.install");
-        }
-        int cmp = com.editora.plugin.PluginInstaller.compareVersions(e.version, installed);
-        return cmp > 0 ? tr("plugins.status.updateAvailable") : tr("plugins.status.installed");
-    }
-
-    /** The installed plugin's manifest version, or null when not installed. */
-    private String installedVersion(String id) {
-        if (pluginManager == null || id == null) {
-            return null;
-        }
-        for (com.editora.plugin.PluginDescriptor d : pluginManager.descriptors()) {
-            if (id.equals(d.id())) {
-                return d.manifest().version == null ? "" : d.manifest().version;
-            }
-        }
-        return null;
-    }
-
-    /** Whether this Editora build satisfies the entry's {@code minEditoraVersion} (blank = any). */
-    private static boolean meetsMinEditora(com.editora.plugin.RegistryEntry e) {
-        String min = e.minEditoraVersion;
-        if (min == null || min.isBlank()) {
-            return true;
-        }
-        return com.editora.plugin.PluginInstaller.compareVersions(com.editora.AppInfo.VERSION, min) >= 0;
-    }
-
-    /** Confirms (showing name/version/author/source + the trust warning) then installs from the registry. */
-    private void confirmAndInstall(com.editora.plugin.RegistryEntry e) {
-        if (e == null) {
-            return;
-        }
-        if (!meetsMinEditora(e)) {
-            setStatus(tr("plugins.status.requiresNewer", e.minEditoraVersion));
-            return;
-        }
-        String body = tr(
-                "dialog.plugins.installBody",
-                (e.name == null || e.name.isBlank() ? e.id : e.name),
-                (e.version == null ? "" : e.version),
-                (e.author == null ? "" : e.author),
-                e.download);
-        if (!browseSigned) {
-            body = tr("dialog.plugins.unsignedWarn") + "\n\n" + body; // reached only when the gate is off
-        }
-        Alert confirm = new Alert(Alert.AlertType.CONFIRMATION, body, ButtonType.OK, ButtonType.CANCEL);
-        confirm.initOwner(stage);
-        confirm.setTitle(tr("dialog.plugins.installTitle"));
-        confirm.setHeaderText(tr("dialog.plugins.installHeader"));
-        if (confirm.showAndWait().orElse(ButtonType.CANCEL) != ButtonType.OK) {
-            return;
-        }
-        setStatus(tr("status.plugins.installing", e.name == null || e.name.isBlank() ? e.id : e.name));
-        pluginInstaller.installFromUrl(e, this::onPluginInstalled);
-    }
-
-    /** Install-from-disk: pick a {@code .zip}, then install it (no checksum — the user chose the file). */
-    private void installPluginFromDisk() {
-        if (!pluginsEnabled() || pluginInstaller == null) {
-            setStatus(tr("status.plugins.disabled"));
-            return;
-        }
-        FileChooser fc = new FileChooser();
-        fc.setTitle(tr("dialog.plugins.installFileTitle"));
-        fc.getExtensionFilters().add(new FileChooser.ExtensionFilter("Plugin zip", "*.zip"));
-        java.io.File f = fc.showOpenDialog(stage);
-        if (f == null) {
-            return;
-        }
-        setStatus(tr("status.plugins.installing", f.getName()));
-        pluginInstaller.installFromZip(f.toPath(), this::onPluginInstalled);
-    }
-
-    /** Common post-install handling: disclose capabilities + confirm enable, persist, refresh, report. */
-    private void onPluginInstalled(com.editora.plugin.PluginInstaller.Result r) {
-        if (!r.ok()) {
-            git.gitError(tr("status.plugins.installFailed", r.error() == null ? "" : r.error()), r.error());
-            return;
-        }
-        pluginRegistry.invalidate(); // status labels (Installed/Update) may change
-        // Arming gate: the plugin is now on disk; show exactly what it can do before enabling it.
-        if (confirmEnablePlugin(r.id())) {
-            config.getPluginStore().setEnabled(r.id(), true);
-            config.savePlugins();
-            setStatus(tr("status.plugins.installed", r.name()));
-        } else {
-            setStatus(tr("status.plugins.notEnabled", r.name()));
-        }
-        settingsWindow.syncPluginsCheck(); // rebuilds the per-plugin list
-    }
-
-    /**
-     * Shows a capability-disclosure confirm before a plugin is <em>enabled</em> (the real arming point —
-     * code loads on next launch): whether it ships executable code, the external commands it declares, and
-     * any keybindings it remaps. Returns whether the user accepted. Falls back to enabling if the descriptor
-     * can't be found.
-     */
-    private boolean confirmEnablePlugin(String id) {
-        com.editora.plugin.PluginDescriptor d = null;
-        if (pluginManager != null) {
-            for (com.editora.plugin.PluginDescriptor c : pluginManager.descriptors()) {
-                if (c.id().equals(id)) {
-                    d = c;
-                    break;
-                }
-            }
-        }
-        if (d == null) {
-            return true;
-        }
-        String name = d.manifest().name == null || d.manifest().name.isBlank() ? d.id() : d.manifest().name;
-        String body = tr(
-                "dialog.plugins.enableBody",
-                name,
-                d.manifest().version == null ? "" : d.manifest().version,
-                pluginCapabilitySummary(d.manifest(), d.hasJavaEntry()));
-        Alert confirm = new Alert(Alert.AlertType.CONFIRMATION, body, ButtonType.OK, ButtonType.CANCEL);
-        confirm.initOwner(stage);
-        confirm.setTitle(tr("dialog.plugins.enableTitle"));
-        confirm.setHeaderText(tr("dialog.plugins.enableHeader"));
-        confirm.getDialogPane().setMinWidth(480);
-        return confirm.showAndWait().orElse(ButtonType.CANCEL) == ButtonType.OK;
-    }
-
-    /**
-     * A human-readable, localized list of what a plugin can do, from its manifest: executable code (a jar),
-     * declared external commands (with their argv), and keybinding remaps. Used by the enable-confirm dialogs
-     * (here and in {@link SettingsWindow}). Pure aside from {@code tr(...)}.
-     */
-    public static String pluginCapabilitySummary(com.editora.plugin.PluginManifest m, boolean hasJar) {
-        StringBuilder sb = new StringBuilder();
-        if (hasJar) {
-            sb.append(tr("plugins.cap.code")).append('\n');
-        }
-        if (m.commands != null && !m.commands.isEmpty()) {
-            sb.append(tr("plugins.cap.commands")).append('\n');
-            for (com.editora.plugin.PluginManifest.DeclaredCommand c : m.commands) {
-                String argv = c.run == null ? "" : String.join(" ", c.run);
-                sb.append("    ").append(argv).append('\n');
-            }
-        }
-        if (m.keymap != null && !m.keymap.isEmpty()) {
-            sb.append(tr("plugins.cap.keymap")).append('\n');
-            m.keymap.forEach((chord, cmd) ->
-                    sb.append("    ").append(chord).append(" → ").append(cmd).append('\n'));
-        }
-        if (sb.length() == 0) {
-            sb.append(tr("plugins.cap.assetsOnly"));
-        }
-        return sb.toString().strip();
-    }
-
-    /** Uninstalls a plugin: deletes its folder + drops it from the enabled store (Settings-page Remove). */
-    private void uninstallPlugin(String id) {
-        if (pluginManager == null || id == null || id.isBlank()) {
-            return;
-        }
-        Alert confirm = new Alert(
-                Alert.AlertType.CONFIRMATION, tr("dialog.plugins.uninstallBody", id), ButtonType.OK, ButtonType.CANCEL);
-        confirm.initOwner(stage);
-        confirm.setTitle(tr("dialog.plugins.uninstallTitle"));
-        confirm.setHeaderText(null);
-        if (confirm.showAndWait().orElse(ButtonType.CANCEL) != ButtonType.OK) {
-            return;
-        }
-        java.nio.file.Path dir = pluginManager.pluginsDir().resolve(id);
-        try (java.util.stream.Stream<java.nio.file.Path> s = java.nio.file.Files.walk(dir)) {
-            s.sorted(java.util.Comparator.reverseOrder()).forEach(p -> {
-                try {
-                    java.nio.file.Files.deleteIfExists(p);
-                } catch (java.io.IOException ignored) {
-                    // best-effort
-                }
-            });
-        } catch (java.io.IOException e) {
-            git.gitError(tr("status.plugins.uninstallFailed", id), e.getMessage());
-            return;
-        }
-        config.getPluginStore().setEnabled(id, false);
-        config.savePlugins();
-        pluginManager.discover();
-        settingsWindow.syncPluginsCheck();
-        setStatus(tr("status.plugins.uninstalled", id));
-    }
-
-    /** Runs a plugin's declared external command via the subprocess runner; reports the result. */
-    private void runDeclaredCommand(
-            com.editora.plugin.PluginDescriptor d, com.editora.plugin.PluginManifest.DeclaredCommand c) {
-        if (c.run == null || c.run.isEmpty()) {
-            return;
-        }
-        java.nio.file.Path cwd = (c.dir == null || c.dir.isBlank())
-                ? d.dir()
-                : d.dir().resolve(c.dir).normalize();
-        setStatus(tr("status.plugins.running", c.title == null || c.title.isBlank() ? c.id : c.title));
-        new Thread(
-                        () -> {
-                            com.editora.process.ProcessRunner.Result r;
-                            try {
-                                r = com.editora.process.ProcessRunner.run(
-                                        cwd,
-                                        java.time.Duration.ofSeconds(120),
-                                        new java.util.ArrayList<>(c.run),
-                                        java.util.Map.of());
-                            } catch (RuntimeException e) {
-                                Platform.runLater(() -> setStatus(tr("status.plugins.cmdFailed", e.getMessage())));
-                                return;
-                            }
-                            String out = (r.out() + "\n" + r.err()).strip();
-                            if (!out.isBlank()) {
-                                LOG.info("[plugin " + d.id() + "] " + out);
-                            }
-                            Platform.runLater(() -> setStatus(
-                                    r.ok()
-                                            ? tr("status.plugins.cmdDone", c.id)
-                                            : tr("status.plugins.cmdFailed", "exit " + r.exit())));
-                        },
-                        "plugin-cmd-" + d.id())
-                .start();
-    }
-
-    /** Builds the plugin-contributed right-click items for {@code buffer} (empty when no plugin added any). */
-    private java.util.List<javafx.scene.control.MenuItem> pluginEditorMenuItems(EditorBuffer buffer) {
-        if (pluginMenuItems.isEmpty()) {
-            return java.util.List.of();
-        }
-        java.util.List<javafx.scene.control.MenuItem> items = new java.util.ArrayList<>();
-        for (EditorMenuContribution c : pluginMenuItems) {
-            javafx.scene.control.MenuItem mi = new javafx.scene.control.MenuItem(c.label());
-            mi.setGraphic(Icons.plugin());
-            mi.setOnAction(e -> {
-                try {
-                    c.action().accept(new ActiveEditorImpl(buffer));
-                } catch (Throwable t) {
-                    LOG.log(java.util.logging.Level.WARNING, "Plugin menu action failed", t);
-                }
-            });
-            items.add(mi);
-        }
-        return items;
-    }
-
-    /** A window-scoped {@link com.editora.plugin.PluginContext}; one per plugin per window. */
-    private final class PluginContextImpl implements com.editora.plugin.PluginContext {
-        private final com.editora.plugin.PluginDescriptor desc;
-
-        PluginContextImpl(com.editora.plugin.PluginDescriptor desc) {
-            this.desc = desc;
-        }
-
-        /** Namespaces a bare command id (no dot) to this plugin; a dotted id (a built-in) is used as-is. */
-        private String fullId(String id) {
-            return id != null && id.indexOf('.') < 0 ? "plugin." + desc.id() + "." + id : id;
-        }
-
-        @Override
-        public void registerCommand(String id, String title, Runnable action) {
-            String cid = fullId(id);
-            registry.register(
-                    com.editora.command.Command.of(cid, title == null || title.isBlank() ? cid : title, action));
-        }
-
-        @Override
-        public void bindKey(String chord, String commandId) {
-            if (chord != null && !chord.isBlank() && commandId != null) {
-                keymap.applyOverrides(java.util.Map.of(chord, fullId(commandId)));
-            }
-        }
-
-        @Override
-        public void registerToolWindow(
-                String id,
-                String title,
-                com.editora.plugin.ToolWindowSide side,
-                javafx.scene.layout.Region content,
-                String commandId,
-                java.util.function.Supplier<javafx.scene.Node> icon,
-                boolean needsBuffer) {
-            String twId = fullId(id);
-            String cmd = commandId == null || commandId.isBlank() ? twId : fullId(commandId);
-            ToolWindow.Side s =
-                    switch (side == null ? com.editora.plugin.ToolWindowSide.BOTTOM : side) {
-                        case LEFT -> ToolWindow.Side.LEFT;
-                        case RIGHT -> ToolWindow.Side.RIGHT;
-                        case BOTTOM -> ToolWindow.Side.BOTTOM;
-                    };
-            // A null supplier (or one returning null) falls back to the default plugin (jigsaw) icon. The
-            // factory is invoked per window/repaint, so it must yield a fresh node each time.
-            java.util.function.Supplier<javafx.scene.Node> iconFactory = icon == null
-                    ? Icons::plugin
-                    : () -> {
-                        javafx.scene.Node n = icon.get();
-                        return n != null ? n : Icons.plugin();
-                    };
-            ToolWindow tw = new ToolWindow(twId, title == null ? twId : title, s, iconFactory, content, cmd);
-            toolWindows.register(tw);
-            if (needsBuffer) {
-                // Acts on the active editor — track it so updateBufferToolWindows() keeps its stripe gated on
-                // an open buffer (hidden on a non-buffer tab like Welcome).
-                pluginToolWindows.add(tw);
-                toolWindows.setAvailable(tw, activeBuffer() != null);
-            }
-            // needsBuffer == false: a self-contained tool window (scratchpad, calculator, …) — never gated.
-            registry.register(
-                    com.editora.command.Command.of(cmd, title == null ? twId : title, () -> toolWindows.toggle(tw)));
-        }
-
-        @Override
-        public void addEditorMenuItem(
-                String label, java.util.function.Consumer<com.editora.plugin.ActiveEditor> action) {
-            if (label != null && action != null) {
-                pluginMenuItems.add(new EditorMenuContribution(label, action));
-            }
-        }
-
-        @Override
-        public void addStatusBarSegment(String label, String commandId) {
-            statusBar.addPluginSegment(label, commandId == null ? null : fullId(commandId));
-        }
-
-        @Override
-        public com.editora.plugin.ActiveEditor activeEditor() {
-            return new ActiveEditorImpl(null); // tracks the live active buffer
-        }
-
-        @Override
-        public java.nio.file.Path pluginDir() {
-            return desc.dir();
-        }
-
-        @Override
-        public java.nio.file.Path dataDir() {
-            java.nio.file.Path data = desc.dir().resolve("data");
-            try {
-                java.nio.file.Files.createDirectories(data);
-            } catch (java.io.IOException ignored) {
-                // best-effort
-            }
-            return data;
-        }
-
-        @Override
-        public java.nio.file.Path configDir() {
-            return config.getConfigDir();
-        }
-
-        @Override
-        public void log(String message) {
-            LOG.info("[plugin " + desc.id() + "] " + message);
-        }
-
-        @Override
-        public void setStatus(String message) {
-            MainController.this.setStatus(message);
-        }
-
-        @Override
-        public void openUrl(String url) {
-            if (url != null && !url.isBlank()) {
-                MainController.this.openExternalUrl(url);
-            }
+        if (pluginCoordinator != null) {
+            pluginCoordinator.disposePlugins();
         }
     }
 
-    /** A {@link com.editora.plugin.ActiveEditor} over a fixed buffer, or the live active buffer when null. */
-    private final class ActiveEditorImpl implements com.editora.plugin.ActiveEditor {
-        private final EditorBuffer fixed;
-
-        ActiveEditorImpl(EditorBuffer fixed) {
-            this.fixed = fixed;
-        }
-
-        private EditorBuffer buf() {
-            return fixed != null ? fixed : activeBuffer();
-        }
-
-        @Override
-        public java.nio.file.Path filePath() {
-            EditorBuffer b = buf();
-            return b == null ? null : b.getPath();
-        }
-
-        @Override
-        public String text() {
-            EditorBuffer b = buf();
-            return b == null ? "" : b.getContent();
-        }
-
-        @Override
-        public String selectedText() {
-            EditorBuffer b = buf();
-            return b == null ? "" : b.getArea().getSelectedText();
-        }
-
-        @Override
-        public int caretLine() {
-            EditorBuffer b = buf();
-            return b == null ? -1 : b.getArea().getCurrentParagraph() + 1; // 1-based
-        }
-
-        @Override
-        public void replaceSelection(String replacement) {
-            EditorBuffer b = buf();
-            if (b != null && b.isEditable() && replacement != null) {
-                b.getArea().replaceSelection(replacement);
+    /** Window hooks for {@link PluginCoordinator} (open-path + the shared Git error dialog). */
+    private PluginCoordinator.Ops pluginOps() {
+        return new PluginCoordinator.Ops() {
+            @Override
+            public void openPath(java.nio.file.Path file) {
+                MainController.this.openPath(file);
             }
-        }
 
-        @Override
-        public void insertAtCaret(String text) {
-            EditorBuffer b = buf();
-            if (b != null && b.isEditable() && text != null) {
-                b.getArea().insertText(b.getArea().getCaretPosition(), text);
+            @Override
+            public void showError(String summary, String detail) {
+                git.gitError(summary, detail);
             }
-        }
-
-        @Override
-        public void setText(String text) {
-            EditorBuffer b = buf();
-            if (b != null && b.isEditable() && text != null) {
-                b.getArea().replaceText(text); // whole-document replace (undoable, marks dirty)
-            }
-        }
-
-        @Override
-        public void openPath(java.nio.file.Path path) {
-            if (path != null) {
-                MainController.this.openPath(path);
-            }
-        }
+        };
     }
 
     /** Shows/hides the toolbar and status bar per the saved settings (hidden nodes also unmanaged so
@@ -1333,9 +789,6 @@ public class MainController implements com.editora.mcp.McpBridge {
         switcher.setOverlayHost(overlayHost);
         branchPopup.setOverlayHost(overlayHost);
         statusBar.setOverlayHost(overlayHost);
-        if (browsePalette != null) {
-            browsePalette.setOverlayHost(overlayHost);
-        }
     }
 
     /**
@@ -1638,16 +1091,6 @@ public class MainController implements com.editora.mcp.McpBridge {
                 ToolWindow::getTitle,
                 tw -> invertBindings().getOrDefault(tw.getCommandId(), ""),
                 toolWindows::open);
-        browsePalette = new QuickOpen<>(
-                tr("plugins.browseTitle"),
-                tr("plugins.browsePrompt"),
-                () -> browseEntries,
-                this::browseEntryLabel,
-                e -> e.description == null ? "" : e.description,
-                e -> (e.name == null ? "" : e.name) + " " + e.id + " "
-                        + (e.description == null ? "" : e.description), // search name+id+desc
-                this::confirmAndInstall);
-        browsePalette.setPreferredSize(960, 10); // wide — registry rows carry a name + a long description
         snippetPalette = new QuickOpen<>(
                 "Insert Snippet",
                 "Type to filter snippets…",
@@ -4796,7 +4239,8 @@ public class MainController implements com.editora.mcp.McpBridge {
         buffer.setOnEnableEditing(() -> enableEditing(buffer)); // "Enable Editing" banner button
         buffer.setSnippetProvider((lang, prefix) -> snippets.byPrefix(lang, prefix));
         buffer.setCompletionProvider(completion::complete);
-        buffer.setMenuContributor(() -> pluginEditorMenuItems(buffer)); // plugin-contributed right-click items
+        buffer.setMenuContributor(
+                () -> pluginCoordinator.editorMenuItems(buffer)); // plugin-contributed right-click items
         Settings acs = config.getSettings();
         buffer.setAutocomplete(
                 acs.isAutocomplete(),
@@ -6376,9 +5820,11 @@ public class MainController implements com.editora.mcp.McpBridge {
         if (externalToolToolWindow != null) {
             toolWindows.setAvailable(externalToolToolWindow, buffer && externalToolsEnabled());
         }
-        // Plugin tool windows act on the active editor — gate them on an open buffer too.
-        for (ToolWindow tw : pluginToolWindows) {
-            toolWindows.setAvailable(tw, buffer);
+        // Plugin tool windows act on the active editor — gate them on an open buffer too. Null-guarded: the
+        // first setupToolWindows() pass runs before pluginCoordinator is built (it needs toolWindows), and no
+        // plugin tool windows exist until applyPlugins() registers them, so an early skip is harmless.
+        if (pluginCoordinator != null) {
+            pluginCoordinator.gateToolWindows(buffer);
         }
         // Git Commit / Git Log act on the active file's repo — hide on a non-buffer tab (e.g. Welcome).
         // When a buffer IS open, leave them to the Git coordinator's in-repo gating (don't force-show).
@@ -6497,14 +5943,6 @@ public class MainController implements com.editora.mcp.McpBridge {
      * so this just persists the preference and reports that a restart is needed; the Settings checkbox is
      * re-synced for discoverability.
      */
-    private void togglePluginSupport() {
-        Settings s = config.getSettings();
-        s.setPluginSupport(!s.isPluginSupport());
-        requestSave();
-        settingsWindow.syncPluginsCheck();
-        setStatus(tr("status.toggle.plugins", tr(s.isPluginSupport() ? "common.on" : "common.off")));
-    }
-
     private void toggleSimpleMode() {
         Settings s = config.getSettings();
         cliSimpleOverride = false; // an explicit in-app toggle takes over from the --simple session flag
@@ -9118,9 +8556,9 @@ public class MainController implements com.editora.mcp.McpBridge {
         registry.register(Command.of("view.toggleColumnRuler", this::toggleColumnRuler));
         registry.register(Command.of("view.toggleToolStripe", this::toggleToolStripe));
         registry.register(Command.of("view.toggleSimpleMode", this::toggleSimpleMode));
-        registry.register(Command.of("view.togglePlugins", this::togglePluginSupport));
-        registry.register(Command.of("plugins.browse", this::browsePlugins));
-        registry.register(Command.of("plugins.installFromDisk", this::installPluginFromDisk));
+        registry.register(Command.of("view.togglePlugins", pluginCoordinator::toggleSupport));
+        registry.register(Command.of("plugins.browse", pluginCoordinator::browse));
+        registry.register(Command.of("plugins.installFromDisk", pluginCoordinator::installFromDisk));
         registry.register(Command.of("config.export", this::exportConfig));
         registry.register(Command.of("editor.setIndentStyle", this::chooseIndentStyle));
         registry.register(Command.of("editor.exportPdf", this::exportCodePdf));
